pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "frameko"
version = "0.1.0"
description = "Notebook-first toolkit: video -> informative frames -> OpenCLIP embeddings -> vector search (FAISS/Qdrant)."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Your Name", email = "you@example.com" }
]
keywords = ["video", "frames", "anime", "clip", "openclip", "faiss", "qdrant", "search"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Multimedia :: Video",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
  "numpy>=1.23",
  "pillow>=10.0",
  "pyyaml>=6.0",
  "tqdm>=4.66",
]

[project.optional-dependencies]
# Embedding
clip = [
  "torch>=2.1",
  "open-clip-torch>=2.24.0",
]
# Scene detection
scenedetect = [
  "scenedetect[opencv]>=0.6.4",
]
# Vector backends
faiss = [
  "faiss-cpu>=1.7.4",
]
qdrant = [
  "qdrant-client>=1.8.0",
]
# Dev tooling
dev = [
  "pytest>=7.4",
  "ruff>=0.5",
  "mypy>=1.7",
  "ipykernel>=6.29",
]

[project.urls]
Homepage = "https://github.com/yourname/frameko"
Repository = "https://github.com/yourname/frameko"
Issues = "https://github.com/yourname/frameko/issues"

[tool.hatch.build]
include = [
  "src/frameko/**",
  "README.md",
  "LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/frameko"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]